[buck] enable junit5
