+++ build.xml
@@ -363,12 +363,17 @@
 
   <fileset dir="${third-party.dir}/java" id="testrunner.jars">
     <include name="android/ddmlib-25.3.0.jar" />
+    <include name="apiguardian-api/apiguardian-api-1.1.2.jar" />
     <include name="guava/guava-20.0.jar" />
     <include name="guava/guava-23.3-jre.jar" />
     <include name="guice/guice-3.0.jar" />
     <include name="jsr/javax.inject-1.jar" />
     <include name="jsr/jsr305.jar" />
     <include name="junit/junit-4.12.jar" />
+    <include name="junit5-jupiter-api/junit-jupiter-api-5.10.1.jar" />
+    <include name="junit5-platform-commons/junit-platform-commons-1.10.1.jar" />
+    <include name="junit5-platform-engine/junit-platform-engine-1.10.1.jar" />
+    <include name="junit5-platform-launcher/junit-platform-launcher-1.10.1.jar" />
     <include name="testng/testng-6.8.jar" />
   </fileset>
 
@@ -1244,7 +1249,12 @@
   <path id="javadoc.classpath">
     <path refid="classpath" />
     <!-- Buck test runner has references to JUnit and TestNG. -->
+    <pathelement location="${third-party.dir}/java/apiguardian-api/apiguardian-api-1.1.2.jar" />
     <pathelement location="${third-party.dir}/java/junit/junit-4.12.jar" />
+    <pathelement location="${third-party.dir}/java/junit5-jupiter-api/junit-jupiter-api-5.10.1.jar" />
+    <pathelement location="${third-party.dir}/java/junit5-platform-commons/junit-platform-commons-1.10.1.jar" />
+    <pathelement location="${third-party.dir}/java/junit5-platform-engine/junit-platform-engine-1.10.1.jar" />
+    <pathelement location="${third-party.dir}/java/junit5-platform-launcher/junit-platform-launcher-1.10.1.jar" />
     <pathelement location="${third-party.dir}/java/testng/testng-6.8.jar" />
 
     <!-- Include the Compiler Tree API stub because its types are used in some places. -->
